remoteproc/pru: switch to client-driven boot methodology
The PRU remoteproc driver currently supports a hybrid boot methodology
- it supports auto-boot only for non PRU-Ethernet firmwares/usecases
and no auto-boot for PRU-Ethernet usecases in which the PRU Ethernet
driver is responsible for booting the PRUs. This is made possible due
to a short-term/temporary solution by using a module parameter and
specific PRU capabilities based on board detection in the driver. This
is not a scalable solution as the driver leveraged specific board
compatibles.

The PRU remoteproc driver is now modified to _not_ support auto-boot
by default for all use-cases so that the PRU load and boot is dictated
by the corresponding client drivers. This allows flexibility for the
client drivers/applications to set a firmware name (if needed) based
on their desired functionality and boot the PRU. The sysfs bind and
unbind attributes have also been suppressed so that the PRU devices
cannot be unbound and thereby shutdown a PRU from underneath a PRU
client driver.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>