net: prueth: add logic to set firmware names
The PRU Ethernet driver has relied so far on the PRU remoteproc
driver having the knowledge about the Ethernet-specific firmware
names. This was previously required to be in the PRU remoteproc
driver on previous kernels because of the two-phase resource table
parsing and loading sequence performed in the remoteproc core.

This sequence is no longer valid on 4.9+ kernels, and the PRU
Ethernet driver has been enhanced to set the firmware name using
the newly added rproc_set_firmware() API before booting a PRU
remoteproc device. This design improves the scalability of both
the PRU remoteproc and PRU Ethernet drivers. The latter driver
can also now be further enhanced to support different Ethernet
protocols using different firmwares.

Note that there is no firmware name context saving in the remoteproc
core, so each application is supposed to configure its firmware
name before booting a PRU remote processor.

Signed-off-by: Suman Anna <s-anna@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>